add more rules, change coins to decimal
