+++ backend/code/archipelag/market/views.py
@@ -7,7 +7,6 @@ from archipelag.market.models import Market
 from archipelag.market.models import POINTS_RULES
 from archipelag.message.models import Message
 from django.http import JsonResponse
-from django.shortcuts import redirect
 from archipelag.ngo.models import NgoUser
 
 
@@ -25,25 +24,29 @@ class MarketView(LoginRequiredMixin, View):
 @login_required
 def market_create(request):
     if request.method == 'POST':
-        body_unicode = request.body.decode('utf-8')
-        body_data = loads(body_unicode)
+        body_data = loads(request.body.decode('utf-8'))
         current_ngo = request.user.ngouser
         if can_create_market(current_ngo):
             delete_points(current_ngo)
-            new_market = Market.objects.create(
-                owner=request.user.ngouser,
-                title=body_data["title"],
-                url=body_data["url"],
-                date_starting=body_data["date_starting"],
-                date_ending=body_data["date_ending"],
-                hashtag=body_data["hashtag"]
-            )
+            new_market = get_new_market(current_ngo, body_data)
             data = dict(url='/message/create/%s'%new_market.id)
             return JsonResponse(data)
         else:
             error = dict(error="Za mało punktów.")
             return JsonResponse(error)
     return render(request, 'registration/event.html', )
+
+
+def get_new_market(current_ngo, body_data):
+    new_market = Market.objects.create(
+        owner=current_ngo,
+        title=body_data["title"],
+        url=body_data["url"],
+        date_starting=body_data["date_starting"],
+        date_ending=body_data["date_ending"],
+        hashtag=body_data["hashtag"]
+    )
+    return new_market
 
 
 def can_create_market(current_ngo):
@@ -64,11 +67,3 @@ def get_messages(request, market_id):
             {
                 'messages': Message.objects.filter(market=market).all(),
             })
-
-
-def add_point(request, market_id):
-    current_ngo = request.user.ngouser
-    ngo_model = NgoUser()
-    ngo_model.add_coins(current_ngo, POINTS_RULES['for_share'])
-    current_ngo.save()
-    return redirect('market_details', market_id=market_id)
